requirements.txt
alembic==0.9.6
aniso8601==1.3.0
astroid==2.4.1
attrs==19.3.0
bcrypt==3.1.4
boto3==1.13.13
botocore==1.16.13
cffi==1.14.0
click==6.7
coverage==4.4.2
docutils==0.15.2
enum-compat==0.0.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.5.3
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.8.0
importlib-metadata==1.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
lazy-object-proxy==1.4.3
Mako==1.0.7
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2==2.8.5
pycparser==2.18
PyJWT==1.7.1
pylint==2.5.2
pyrsistent==0.16.0
python-dateutil==2.6.1
python-dotenv==0.13.0
python-editor==1.0.3
python-magic==0.4.18
pytz==2020.1
s3transfer==0.3.3
six==1.14.0
SQLAlchemy==1.3.17
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
Werkzeug==0.16.1
wrapt==1.12.1
zipp==3.1.0
